# docs/openapi-banco-orbital.yaml
...
openapi: 3.0.3
info:
  title: Banco Orbital — API
  version: 1.0.0-draft
  description: |
    API del home banking de **Banco Orbital**. Este documento es el contrato entre
    nuestro backend y nuestro frontend, y la referencia para los otros bancos del
    curso que quieran comparar.

    ## Alcance de la v1

    Cuentas, tarjetas, préstamos y plazos fijos. **Seguros, CEDEARs y reservas
    quedan para la v2** — decisión tomada para que el alcance entre en el tiempo
    disponible.

    Los proveedores externos simulados (recargas de celular y empresas de
    servicios) **no están acá**: viven en una API aparte, `banco-proveedores`, que
    este banco consume por HTTP como si fuera un tercero real.

    ## Convenciones

    | Tema | Convención |
    |---|---|
    | Autenticación | JWT de Clerk en `Authorization: Bearer <token>` |
    | Errores | `{ "error": "mensaje" }`, con `details` opcional. Igual que el Banco Central |
    | Paginación | `page` y `limit` por query; respuesta `{ page, limit, count, data }` |
    | Importes | `number` con 2 decimales. Internamente se calculan en decimal exacto |
    | Nombres | Español, según `docs/GLOSARIO.md`. Campos en `snake_case`, como ya los devuelven los DTOs |
    | Idempotencia | Header `Idempotency-Key` (UUID v4) en **todo endpoint que mueve plata** |

    ### Sobre `monto` contra `importe`

    Dentro de nuestro banco el campo siempre se llama **`monto`**. `importe` aparece
...
    contrato. No mezclarlos.

    ### Sobre las rutas
...
servers:
...
security:
  - ClerkJWT: []
...
components:
  securitySchemes:
    ClerkJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token de sesión emitido por Clerk. El backend lo valida y resuelve el
        `usuario` y su `persona` asociada. Todos los endpoints bajo `/api` lo
        requieren, además de usuario activo y perfil completo.
  parameters:
    Page:
      name: page
      in: query
      required: false
      schema: { type: integer, minimum: 1, default: 1 }
      description: Página, empezando en 1.
    Limit:
      name: limit
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      description: Cantidad de elementos por página.
...
  responses:
    NoAutorizado:
      description: Falta el token, es inválido, o el usuario está inactivo.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: 'No autorizado.' }
    Prohibido:
      description: El usuario no tiene permiso, o el recurso pertenece a otra persona.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: 'No puedes operar sobre una cuenta que no te pertenece.' }
    NoEncontrado:
      description: El recurso no existe.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: 'No existe el recurso solicitado.' }
    ErrorValidacion:
      description: El cuerpo o los parámetros no pasaron la validación.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error: 'Datos inválidos.'
            details:
              fieldErrors:
                monto: ['El monto debe ser mayor a cero.']
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Mensaje legible. En producción los 5xx se reemplazan por uno genérico.
          example: 'Saldo insuficiente.'
        details:
          type: object
          nullable: true
          description: |
            Detalle adicional. Presente en errores de validación (con `fieldErrors`)
            y en desarrollo. Nunca expone internos de la base en producción.
      additionalProperties: false
    Paginado:
      type: object
      description: Envoltorio de todos los listados paginados.
      required: [page, limit, count, data]
      properties:
        page: { type: integer, example: 1 }
        limit: { type: integer, example: 20 }
        count:
          type: integer
          description: Cantidad de elementos en **esta** página, no el total del recurso.
          example: 20
        data:
          type: array
          items: { type: object }
    Moneda:
      type: string
      enum: [ARS, USD]
      description: Moneda de la cuenta. La caja en pesos nace con la persona; la de dólares se abre aparte.
      example: ARS
    Monto:
      type: number
      format: double
      description: |
        Importe con 2 decimales. Se almacena como `NUMERIC(18,2)` y se calcula en
        decimal exacto; se expone como número por comodidad del cliente, que sólo
        lo formatea para mostrar.
      example: 184500.20
    Cbu:
      type: string
      minLength: 22
      maxLength: 22
      pattern: '^\d{22}$'
      description: Clave Bancaria Uniforme. Cada cuenta tiene el suyo, así que la caja en pesos y la de dólares de una misma persona tienen CBU distinto.
      example: '0060001948123456001608'
    Cuenta:
      type: object
      properties:
        id: { type: string, format: uuid }
        persona_id: { type: string, format: uuid }
        tipo_cuenta_id: { type: string, format: uuid }
        numero_cuenta: { type: string, example: '001948123456' }
        cbu: { $ref: '#/components/schemas/Cbu' }
        alias:
          type: string
          nullable: true
          example: juan.perez.orbital
        moneda: { $ref: '#/components/schemas/Moneda' }
        saldo: { $ref: '#/components/schemas/Monto' }
        principal:
          type: boolean
          description: |
            La cuenta que nació junto con la persona, siempre en pesos. Es la que el
            Banco Central usa como identidad del titular en sus endpoints heredados.
            Hay exactamente una por persona.
          example: true
        activa: { type: boolean, example: true }
        banco_central_registrada:
          type: boolean
          description: Si la cuenta ya está dada de alta en el Banco Central y puede operar entre bancos.
          example: true
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        tipo_cuenta_nombre:
          type: string
          description: Sólo viene cuando la query hizo el JOIN.
          example: Caja de Ahorro
    Transaccion:
      type: object
      properties:
        id: { type: string, format: uuid }
        tipo_transaccion_id: { type: string, format: uuid }
        cuenta_origen_id:
          type: string
          format: uuid
          nullable: true
          description: '`null` en un depósito en efectivo: el dinero no sale de ninguna cuenta.'
        cuenta_destino_id:
          type: string
          format: uuid
          nullable: true
          description: '`null` cuando el destino es de otro banco.'
        cbu_origen: { type: string, nullable: true }
        cbu_destino: { type: string, nullable: true }
        monto: { $ref: '#/components/schemas/Monto' }
        moneda: { $ref: '#/components/schemas/Moneda' }
        descripcion: { type: string, nullable: true, example: Pago alquiler }
        estado:
          type: string
          enum: [pendiente, completada, rechazada]
          example: completada
        canal:
          type: string
          nullable: true
          enum: [local, interbancaria_saliente, interbancaria_entrante, deposito_efectivo, extraccion_efectivo, cambio_divisa]
          description: Por dónde pasó la operación. Lo usa el frontend para elegir el ícono y el texto.
          example: local
        central_transaction_id:
          type: string
          nullable: true
          description: Identificador que devolvió el Banco Central, si la operación pasó por él.
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    Cotizacion:
      type: object
      description: Cotización tomada de DolarAPI, casa `oficial`, cacheada unos minutos.
      properties:
        moneda: { type: string, example: USD }
        casa: { type: string, example: oficial }
        compra:
          type: number
          description: Pesos que el banco paga por cada dólar que le compra al cliente.
          example: 1465
        venta:
          type: number
          description: Pesos que el cliente paga por cada dólar que le compra al banco.
          example: 1515
        fecha_actualizacion: { type: string, format: date-time }
    Tarjeta:
      type: object
      properties:
        id: { type: string, format: uuid }
        persona_id: { type: string, format: uuid }
        tipo:
          type: string
          enum: [debito, credito]
          example: credito
        numero_enmascarado:
          type: string
          description: Sólo los últimos 4 dígitos. El número completo no se expone nunca.
          example: '**** **** **** 4821'
        cuenta_id:
          type: string
          format: uuid
          nullable: true
          description: Cuenta asociada. Obligatoria en débito, `null` en crédito.
        limite:
          allOf: [{ $ref: '#/components/schemas/Monto' }]
          nullable: true
          description: Sólo en crédito.
        disponible:
          allOf: [{ $ref: '#/components/schemas/Monto' }]
          nullable: true
          description: 'Límite menos el consumo del período. Sólo en crédito.'
        estado:
          type: string
          enum: [activa, bloqueada, vencida]
          example: activa
        vencimiento: { type: string, example: '09/30' }
        created_at: { type: string, format: date-time }
...
paths:
...
  /api/personas/{personaId}/cuentas:
    get:
      tags: [Cuentas]
      summary: Listar las cuentas de una persona
      description: |
        Devuelve todas las cuentas del titular, en pesos y en dólares. Es la consulta
        que arma el resumen del portal.

        Un cliente sólo puede ver las suyas; los roles internos pueden ver las de
        cualquiera.
      parameters:
        - name: personaId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Lista de cuentas
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Cuenta' }
        '401': { $ref: '#/components/responses/NoAutorizado' }
        '403': { $ref: '#/components/responses/Prohibido' }
        '404': { $ref: '#/components/responses/NoEncontrado' }
  /api/personas/{personaId}/cuentas/apertura:
    post:
      tags: [Cuentas]
      summary: Abrir una caja de ahorro
      description: |
        Abre una caja de ahorro en la moneda pedida y la registra en el Banco Central
        con `POST /accounts`, guardando el CBU que devuelve.

        **Por qué se guarda el CBU localmente:** el Banco Central no tiene forma de
        listar las cuentas de una persona, así que si lo perdemos el único modo de
        recuperarlo es volver a llamar a `POST /accounts` con el mismo DNI y moneda.

        Antes de abrir se consulta la central de deudores: **situación 3 o peor
        rechaza la apertura**.
      parameters:
        - name: personaId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [moneda]
              properties:
                moneda: { $ref: '#/components/schemas/Moneda' }
                alias:
                  type: string
                  nullable: true
                  description: Si se omite, se genera uno a partir del nombre.
                  example: juan.perez.usd
      responses:
        '201':
          description: Cuenta creada y registrada en el Banco Central.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Cuenta' }
        '200':
          description: La persona ya tenía una cuenta en esa moneda. Se devuelve la existente.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Cuenta' }
        '400': { $ref: '#/components/responses/ErrorValidacion' }
        '401': { $ref: '#/components/responses/NoAutorizado' }
        '403':
          description: |
            Sin permiso sobre esa persona, **o** el titular está en situación 3 o peor
            en la central de deudores.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example:
                error: 'No se puede abrir la cuenta: el titular registra situación 4 en la central de deudores.'
        '404': { $ref: '#/components/responses/NoEncontrado' }
  /api/cuentas/{cuentaId}/movimientos:
    get:
      tags: [Cuentas]
      summary: Movimientos de una cuenta
      description: Extracto paginado, del más reciente al más antiguo.
      parameters:
        - name: cuentaId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Movimientos de la cuenta
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Paginado'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/Transaccion' }
        '401': { $ref: '#/components/responses/NoAutorizado' }
        '403': { $ref: '#/components/responses/Prohibido' }
        '404': { $ref: '#/components/responses/NoEncontrado' }

  # ──────────────────────────── Transacciones ────────────────────────────
...
  /api/personas/{personaId}/tarjetas:
    get:
      tags: [Tarjetas]
      summary: Listar las tarjetas de una persona
      parameters:
        - name: personaId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Tarjetas del titular.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Tarjeta' }
        '401': { $ref: '#/components/responses/NoAutorizado' }
        '403': { $ref: '#/components/responses/Prohibido' }
...
  /api/catalogos/cotizacion:
    get:
      tags: [Catálogos]
      summary: Cotización del dólar
      description: |
        Cotización vigente que usa el banco para el cambio de divisa. Sale de DolarAPI,
        casa `oficial`, con caché de unos minutos y valor de respaldo si la API externa
        se cae.
      responses:
        '200':
          description: Cotización vigente.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Cotizacion' }
        '401': { $ref: '#/components/responses/NoAutorizado' }
        '503':
          description: La API externa no responde y no hay valor cacheado.
  /api/catalogos/tasas:
    get:
      tags: [Catálogos]
      summary: Tasas de referencia
      description: |
        Tasas que el banco ofrece, derivadas de ArgentinaDatos.

        **Ojo:** ArgentinaDatos publica la TNA como fracción decimal (`0.19` para
        19 %). Acá se devuelve ya normalizada **en porcentaje**, que es como la
        esperan los cálculos.
      responses:
        '200':
          description: Tasas vigentes.
          content:
            application/json:
              schema:
                type: object
                properties:
                  prestamos:
                    type: object
                    properties:
                      tna: { type: number, example: 72 }
                  plazo_fijo:
                    type: object
                    properties:
                      tna: { type: number, example: 58 }
                  actualizado_al: { type: string, format: date-time }
        '401': { $ref: '#/components/responses/NoAutorizado' }